requirements.txt
pyyaml
tqdm
pandas
numpy
seaborn
torch
torchvision
scikit-learn